+++ src/duqtools/jetto/__init__.py
@@ -1,7 +1,7 @@
 """This module contains tools for interfacing with jetto runs."""
 
+from ._batchfile import write_batchfile
 from ._jettovar_to_json import jettovar_to_json
-from ._llcmd import write_batchfile
 from ._system import BaseJettoSystem, JettoSystem, JettoSystemV210921, JettoSystemV220922
 
 __all__ = [
