+++ src/duqtools/jetto/_batchfile.py
